requirements.txt
numpy==1.20.3
opencv-python==4.10.0.84
pillow==10.4.0
shapely==2.0.7